Refactor JSON functionality into its own package. JsonRequestCall is the base abstract class for all HTTPS JSON-based requests. JsonApiCall subclasses JSONRequestCall for use with the JsonApiCondition. Update imports and references accordingly.
