+++ nucypher/policy/conditions/lingo.py
@@ -697,7 +697,7 @@ class ConditionLingo(_Serializable):
         conditions expression framework.
         """
         from nucypher.policy.conditions.evm import ContractCondition, RPCCondition
-        from nucypher.policy.conditions.offchain import JsonApiCondition
+        from nucypher.policy.conditions.json.api import JsonApiCondition
         from nucypher.policy.conditions.time import TimeCondition
 
         # version logical adjustments can be made here as required
